% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexr.R
\name{latex2r}
\alias{latex2r}
\title{Translate LaTeX formula to R code.}
\usage{
latex2r(x, interactive = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{string}

\item{interactive}{boolean}

\item{na.rm}{boolean; when \code{TRUE}, calls to \code{mean()} and \code{median()} are
emitted with \code{na.rm = TRUE}.}
}
\value{
string
}
\description{
Take a LaTeX formula and attempts to find an equivalent representation in R.
If \code{interactive=TRUE} it pops up an interactive REPL that does the same, but interactively.
}
\examples{
latex2r("\\\\beta_1^{\\\\frac{x+1}{x^2 \\\\cdot y}}")
latex2r("\\\\bar{x}", na.rm = TRUE)
}
