% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex-input.R
\name{latexInput}
\alias{latexInput}
\title{Latex Input}
\usage{
latexInput(inputId, label, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};}
}
\value{
A latex input control that can be added to a UI definition.
}
\description{
Latex Input
}
