% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-app.R
\name{launch_app}
\alias{launch_app}
\title{Run an interactive LaTeX to R translator which uses \code{\link[=latexInput]{latexInput()}}}
\usage{
launch_app()
}
\description{
Run an interactive LaTeX to R translator which uses \code{\link[=latexInput]{latexInput()}}
}
