% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Token.R
\name{print.Token}
\alias{print.Token}
\title{Print a Token object.}
\usage{
\method{print}{Token}(x, ...)
}
\arguments{
\item{x}{A \code{Token} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The string representation of the token, invisibly.
}
\description{
Print a Token object.
}
