% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex2fun.R
\name{latex2fun}
\alias{latex2fun}
\title{Construct R function from LaTeX expression.}
\usage{
latex2fun(latex_string, envir = parent.frame())
}
\arguments{
\item{latex_string}{string}

\item{envir}{environment}
}
\value{
function
}
\description{
Construct R function from LaTeX expression.
}
\examples{
x = seq(-2 * pi, 2 * pi, length.out = 500)
f = latex2fun("\\\\sin{x * a}")
f_x = f(x = x, a = 2)
plot(x, f_x, type = "l")
}
